refactor: change new project folder structure
