+++ client-side/src/pages/Analysis/Analysis.jsx
@@ -1,7 +1,7 @@
-import Aside from "../Aside/Aside"
-import Navbar from "../Navbar/Navbar"
+import Aside from "../../components/Aside/Aside"
+import Navbar from "../../components/Navbar/Navbar"
 
-function AddStock() {
+function Analysis() {
   return (
     <div className=" w-full h-[100vh]">
     <div className="w-full absolute">
@@ -17,12 +17,11 @@ function AddStock() {
         className="w-full bg-[#F9FAFB] h-[100vh] mt-[3rem]"
       >
         <div className="border w-full h-full">
-          Add stock
+          Analysis
         </div>
       </div>
     </div>
   </div>
   )
 }
-
-export default AddStock
+export default Analysis;
