+++ client-side/src/pages/Sale/ProductNoBarcode.jsx
