+++ client-side/src/pages/Sale/Scan.jsx
