+++ client-side/src/pages/Stock/Stock.jsx
@@ -1,7 +1,7 @@
-import Aside from "../Aside/Aside"
-import Navbar from "../Navbar/Navbar"
+import Aside from "../../components/Aside/Aside"
+import Navbar from "../../components/Navbar/Navbar"
 
-function Analysis() {
+function Stock() {
   return (
     <div className=" w-full h-[100vh]">
     <div className="w-full absolute">
@@ -17,11 +17,12 @@ function Analysis() {
         className="w-full bg-[#F9FAFB] h-[100vh] mt-[3rem]"
       >
         <div className="border w-full h-full">
-          Analysis
+          Stock
         </div>
       </div>
     </div>
   </div>
   )
 }
-export default Analysis;
+
+export default Stock;
